Initial commit
It's a big one, because I had to get this thing working in principle.
And most of the volume comes simply from importing the old site.

I deleted some of the conversion scripts to not run them again by
accident and overwrite my manual adaptions, but now I think that was
very weird and kind of stupid. Let's hope I really don't need them
again.