Update htcpp.joml with new htcpp features
This includes environment variable expansion, which is very helpful and
avoids me forgetting to change the port before I run locally or deploy.

It also includes headers, so I can add a C ache-Control header. And last
but not least a huge number of routes could be replaced with a simple
pattern.